My Delegation being properly displayed (for fixture data only at this point)
