+++ packages/webapp/src/elections/components/election-member-chips.tsx
@@ -1,7 +1,7 @@
 import { useRouter } from "next/router";
 import { FaCheckSquare, FaPlayCircle, FaRegSquare } from "react-icons/fa";
 
-import { ipfsUrl, openInNewTab } from "_app";
+import { ipfsUrl, openInNewTab, useElectionState, useHeadDelegate } from "_app";
 import { ROUTES } from "_app/config";
 import { GenericMemberChip } from "_app/ui";
 import { MemberData } from "members/interfaces";
@@ -80,11 +80,29 @@ export const VotingMemberChip = ({
 
 interface DelegateChipProps {
     member?: MemberData;
-    level?: string;
+    level?: number;
 }
 
+const getDelegateLevelDescription = (
+    memberAccount: string | undefined,
+    level: number | undefined
+) => {
+    if (!memberAccount || !level) return "Delegate";
+    const prefix = `D${level} - `;
+    const { data: leadRepresentative } = useHeadDelegate();
+    const { data: electionState } = useElectionState();
+    if (leadRepresentative === memberAccount) return prefix + "Head Chief";
+    if (electionState?.board.includes(memberAccount))
+        return prefix + "Chief Delegate";
+    return "Delegate";
+};
+
 export const DelegateChip = ({ member, level }: DelegateChipProps) => (
-    <ElectionParticipantChip member={member} delegateLevel={level} isDelegate />
+    <ElectionParticipantChip
+        member={member}
+        delegateLevel={getDelegateLevelDescription(member?.account, level)}
+        isDelegate
+    />
 );
 
 interface ElectionParticipantChipProps {
